Simplify AwtGameWindow*
